Implement aes128gcm, remove legacy code, add exactor agnostic client, update docs, and more. (#30)
* WIP : implementing ECE with ECE crate. Missing tests. Not compiling yet

* Restaured front padding for both encryption methods

* Removed structures that are no longer in use

* Removed tests that are now handled by the rust_ece crate

* Removed obsolete comment

* Updated readme to include new standard

* Removed tests that are covered in the rust-ece crate
prepared tests for headers that are left to do

* Implemented some of the headers tests

* Fixed numbers with new double-padding
(some small padding is added in the rust-ece crate)

* Implemented last tests for aes128gcm

* Cargo fmt

* WIP : added option for testing in the example file

* Cargo fmt

* Remove padding from aes128Gcm.
This allows for aes128Gcm to fully operate in all tested browsers, including edge.

* Remove unneeded base64URL encoding.

* Remove uses of aesgcm.

* Fix tests, store VAPID keys as bytes, and add reusable VAPID builder.
This commit will fix the base64 encoding issues by directly reading the PEM key as bytes instead of immediately encoding, allowing consumers to choose to base64 encode or not.

* Fix doc-tests.

* Remove FCM specific code.
Use the autopush implementation for all connections. This is possible as web push has been standardised, leaving this legacy code redundant. The FCM crate should cover all uses of FCM by this crates consumers.

* Use isahc as the default client backend.
Add `hyper-client` feature to optionally use the old hyper client. Isach is runtime independent, allowing this crate to be used in non Tokio environments. The hyper client remains as a potentially faster and more maintained option. Isach is the default as I feel the crate should work out of the box on any async executor.

* Update docs and README.md.
Expose request_builder now that it's generic and documented.

* Bump to v0.8

* Add method to get the public key from VAPID signature.

* Add encryption test and change fake subs into real subs.

* Fix final doc test.

* Add migration info to README.md.

Co-authored-by: John Tiesselune <jsos10@pm.me>